Add `hidden` support to contents

+++ api/src/favorites/favorites.service.ts
@@ -3,6 +3,9 @@ import { InjectRepository } from '@mikro-orm/nestjs';
 import { BadRequestException, Injectable } from '@nestjs/common';
 import { Content } from '../contents/content.entity';
 import { BaseRepository } from '../shared/lib/repositories/base.repository';
+import { assertPermissions } from '../shared/lib/utils/assert-permission';
+import { Action } from '../shared/modules/authorization';
+import { CaslAbilityFactory } from '../shared/modules/casl/casl-ability.factory';
 import type { PaginationOptions } from '../shared/modules/pagination/pagination-option.interface';
 import type { PaginatedResult } from '../shared/modules/pagination/pagination.interface';
 import type { User } from '../users/user.entity';
@@ -13,10 +16,14 @@ export class FavoritesService {
   constructor(
     @InjectRepository(Favorite) private readonly favoriteRepository: BaseRepository<Favorite>,
     @InjectRepository(Content) private readonly contentRepository: BaseRepository<Content>,
+    private readonly caslAbilityFactory: CaslAbilityFactory,
   ) {}
 
   public async create(contentId: number, user: User): Promise<Favorite> {
     const content = await this.contentRepository.findOneOrFail({ contentId }, { populate: ['parent'] });
+
+    const ability = this.caslAbilityFactory.createForUser(user);
+    assertPermissions(ability, Action.Interact, content);
 
     const favorite = new Favorite({ content, user });
     try {
@@ -30,9 +37,11 @@ export class FavoritesService {
   }
 
   public async findAll(user: User, paginationOptions?: PaginationOptions): Promise<PaginatedResult<Favorite>> {
+    const canSeeHiddenContent = this.caslAbilityFactory.canSeeHiddenContent(user);
+    const visibilityQuery = canSeeHiddenContent ? {} : { content: { isVisible: true } };
     return await this.favoriteRepository.findWithPagination(
       paginationOptions,
-      { user },
+      { user, ...visibilityQuery },
       {
         populate: [
           'user',
@@ -54,11 +63,21 @@ export class FavoritesService {
         ],
       },
     );
+
+    if (favorite) {
+      const ability = this.caslAbilityFactory.createForUser(user);
+      assertPermissions(ability, Action.Read, favorite.content);
+    }
+
     return favorite;
   }
 
   public async remove(contentId: number, user: User): Promise<void> {
     const favorite = await this.favoriteRepository.findOneOrFail({ content: { contentId }, user });
+
+    const ability = this.caslAbilityFactory.createForUser(user);
+    assertPermissions(ability, Action.Interact, favorite.content);
+
     await this.favoriteRepository.removeAndFlush(favorite);
   }
 }
